Revert "Settings: Move Notch Cutout to Display"
This is reverted because notch cutout options are NOT visible in display, without this reverted they are removed from the rom. This reverts add them back to development options so we can use them

This reverts commit a12fe98c3e4fa561b9b59924ae24b01415618677.
