+++ res/xml/development_settings.xml
@@ -469,6 +469,11 @@
             android:key="density"
             android:title="@string/developer_smallest_width" /-->
 
+        <ListPreference
+            android:key="display_cutout_emulation"
+            android:title="@string/display_cutout_emulation"
+            settings:keywords="@string/display_cutout_emulation_keywords" />
+
     </PreferenceCategory>
 
     <PreferenceCategory
