+++ src/com/android/settings/development/EmulateDisplayCutoutPreferenceController.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.android.settings.display;
+package com.android.settings.development;
 
 import android.content.Context;
 import android.content.om.IOverlayManager;
